requirements.txt
Django>=4.2
opencv-python-headless>=4.7.0
Pillow>=10.0
numpy>=1.24
djangorestframework>=3.14
python-multipart>=0.0.6
